+++ src/helper.js
@@ -38,10 +38,10 @@ class Helper {
 
     /**
      * Returns information regarding the current state of the test run
-     * @returns {{testCases: TestCase[], currentFeature: Feature, currentTestCase: TestCase, currentStep: Step}} A clone of the current run state
+     * @returns {HelperRunState} A clone of the current run state
      */
     get currentRun() {
-        return require('clone-deep')(singleton.core.currentRun);
+        return require('safe-clone-deep')(singleton.core.currentRun);
     }
 
     /**